+++ hydro_lang/Cargo.toml
@@ -26,17 +26,21 @@ deploy = [
 ]
 build = ["dep:dfir_lang"]
 runtime_measure = ["dep:chrono", "dep:procfs"]
+staged_macro = ["dep:dfir_rs"]
+runtime_support = ["dep:dfir_rs", "dfir_rs/deploy_integration"]
+dfir_context = ["dep:dfir_rs"]
 
 [package.metadata.docs.rs]
 all-features = true
 
 [dependencies]
+bytes = "1.1.0"
 bincode = "1.3.1"
 hydro_deploy = { path = "../hydro_deploy/core", version = "^0.12.1", optional = true }
-dfir_rs = { path = "../dfir_rs", version = "^0.12.1", default-features = false, features = [
-    "deploy_integration",
-] }
+hydro_deploy_integration = { path = "../hydro_deploy/hydro_deploy_integration", version = "^0.12.1" }
+dfir_rs = { path = "../dfir_rs", version = "^0.12.1", default-features = false, optional = true }
 dfir_lang = { path = "../dfir_lang", version = "^0.12.1", optional = true }
+futures = "0.3.0"
 match_box = "0.0.2"
 nameof = { version = "1.0.0", optional = true }
 prettyplease = { version = "0.2.0", features = ["verbatim"], optional = true }
@@ -55,7 +59,9 @@ syn = { version = "2.0.46", features = [
     "visit-mut",
 ] }
 tokio = { version = "1.29.0", features = ["full"] }
-tokio-stream = { version = "0.1.3", default-features = false }
+tokio-stream = { version = "0.1.3", default-features = false, features = [
+    "time",
+] }
 toml = { version = "0.8.0", optional = true }
 trybuild-internals-api = { version = "1.0.99", optional = true }
