+++ hydro_lang/src/lib.rs
@@ -2,16 +2,28 @@
 
 stageleft::stageleft_no_entry_crate!();
 
-pub use dfir_rs;
 pub use stageleft::q;
 
+#[cfg(feature = "runtime_support")]
+#[cfg_attr(docsrs, doc(cfg(feature = "runtime_support")))]
 #[doc(hidden)]
 pub mod runtime_support {
-    pub use {bincode, stageleft, tokio};
+    pub use {bincode, dfir_rs, stageleft, tokio};
     pub mod resource_measurement;
 }
 
+#[doc(hidden)]
+pub mod internal_constants {
+    pub const CPU_USAGE_PREFIX: &str = "CPU:";
+    // Should remain consistent with dfir_lang/src/graph/ops/_counter.rs
+    pub const COUNTER_PREFIX: &str = "_counter";
+}
+
+#[cfg(feature = "dfir_context")]
+#[cfg_attr(docsrs, doc(cfg(feature = "dfir_context")))]
 pub mod runtime_context;
+#[cfg(feature = "dfir_context")]
+#[cfg_attr(docsrs, doc(cfg(feature = "dfir_context")))]
 pub use runtime_context::RUNTIME_CONTEXT;
 
 pub mod boundedness;
