+++ hydro_lang/src/location/mod.rs
@@ -2,7 +2,7 @@ use std::fmt::Debug;
 use std::marker::PhantomData;
 use std::time::Duration;
 
-use dfir_rs::futures::stream::Stream as FuturesStream;
+use futures::stream::Stream as FuturesStream;
 use proc_macro2::Span;
 use stageleft::{QuotedWithContext, q};
