+++ hydro_std/Cargo.toml
@@ -17,6 +17,7 @@ all-features = true
 [dependencies]
 hydro_lang = { path = "../hydro_lang", version = "^0.12.1" }
 stageleft = "0.8.1"
+serde = "1"
 hdrhistogram = "7.5.4"
 stats-ci = "0.1.1"
