+++ hydro_test_local/src/local/compute_pi.rs
@@ -1,7 +1,7 @@
 use std::time::Duration;
 
+use dfir_rs::scheduled::graph::Dfir;
 use hydro_lang::deploy::SingleProcessGraph;
-use hydro_lang::dfir_rs::scheduled::graph::Dfir;
 use hydro_lang::*;
 use stageleft::{Quoted, RuntimeData};
