+++ hydro_test_local/src/local/first_ten.rs
@@ -1,5 +1,5 @@
+use dfir_rs::scheduled::graph::Dfir;
 use hydro_lang::deploy::SingleProcessGraph;
-use hydro_lang::dfir_rs::scheduled::graph::Dfir;
 use hydro_lang::*;
 use stageleft::Quoted;
