+++ template/hydro/Cargo.toml
@@ -25,3 +25,4 @@ hydro_lang = { git = "{{ hydro_git | default: 'https://github.com/hydro-project/
 ] }
 tokio = { version = "1.29.0", features = ["full"] }
 tokio-stream = { version = "0.1.3", default-features = false }
+futures = "0.3.0"
