+++ template/hydro/src/first_ten_cluster.rs
@@ -15,9 +15,9 @@ pub fn first_ten_cluster<'a>(leader: &Process<'a, Leader>, workers: &Cluster<'a,
 
 #[cfg(test)]
 mod tests {
+    use futures::StreamExt;
     use hydro_deploy::Deployment;
     use hydro_lang::deploy::DeployCrateWrapper;
-    use hydro_lang::dfir_rs::futures::StreamExt;
     use tokio_stream::wrappers::UnboundedReceiverStream;
 
     #[tokio::test]
